recipe_buddy_system/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
spacy==3.7.2
pyspellchecker==0.8.1
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
